server/pyproject.toml
[tool.poetry]
name = "loki-controller-server"
version = "1.0.0"
description = "WebSocket server that allows Loki to be controlled by an accompaying webapp"
authors = ["PurpleMyst <PurpleMyst@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
websockets = "^13.1"
pyserial = "^3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
